feat(DCOS-15159): Implement delete service confirmation field

+++ plugins/services/src/js/components/modals/ServiceDestroyModal.js
@@ -1,11 +1,13 @@
-import { Confirm } from "reactjs-components";
+import { Confirm, Modal } from "reactjs-components";
 import { routerShape } from "react-router";
 import PureRender from "react-addons-pure-render-mixin";
 import React, { PropTypes } from "react";
+import { injectIntl, intlShape } from "react-intl";
 
 import ModalHeading from "#SRC/js/components/modals/ModalHeading";
 import StringUtil from "#SRC/js/utils/StringUtil";
 import UserActions from "#SRC/js/constants/UserActions";
+import ClickToSelect from "#SRC/js/components/ClickToSelect";
 
 import AppLockedMessage from "./AppLockedMessage";
 import Framework from "../../structs/Framework";
@@ -15,14 +17,19 @@ import ServiceTree from "../../structs/ServiceTree";
 
 // This needs to be at least equal to @modal-animation-duration
 const REDIRECT_DELAY = 300;
-const METHODS_TO_BIND = ["handleRightButtonClick"];
+const METHODS_TO_BIND = [
+  "handleChangeInputFieldDestroy",
+  "handleModalClose",
+  "handleRightButtonClick"
+];
 
 class ServiceDestroyModal extends React.Component {
   constructor() {
     super(...arguments);
 
     this.state = {
-      errorMsg: null
+      errorMsg: null,
+      serviceNameConfirmationValue: ""
     };
 
     this.shouldComponentUpdate = PureRender.shouldComponentUpdate.bind(this);
@@ -76,41 +83,27 @@ class ServiceDestroyModal extends React.Component {
     return this.state.errorMsg && /force=true/.test(this.state.errorMsg);
   }
 
-  handleRightButtonClick() {
-    this.props.deleteItem(this.shouldForceUpdate());
+  handleModalClose() {
+    this.setState({ serviceNameConfirmationValue: "" });
+    this.props.onClose();
   }
 
-  getDestroyMessage() {
-    const { service } = this.props;
-    let serviceName = "";
-
-    if (service) {
-      serviceName = service.getId();
+  handleRightButtonClick() {
+    if (!this.getIsRightButtonDisabled()) {
+      this.props.deleteItem(this.shouldForceUpdate());
+      this.setState({ serviceNameConfirmationValue: "" });
     }
+  }
 
-    if (service instanceof Framework) {
-      return (
-        <p>
-          This will only
-          {" "}
-          {UserActions.DELETE}
-          {" "}
-          the package scheduler for
-          {" "}
-          <span className="emphasize">{serviceName}</span>
-          . Any tasks that were started by this scheduler will persist in the system. Are you sure you want to continue?
-        </p>
-      );
-    }
+  handleChangeInputFieldDestroy(event) {
+    this.setState({
+      serviceNameConfirmationValue: event.target.value
+    });
+  }
 
+  getIsRightButtonDisabled() {
     return (
-      <p>
-        {StringUtil.capitalize(UserActions.DELETING)}
-        {" "}
-        <span className="emphasize">{serviceName}</span>
-        {" "}
-        is irreversible. Are you sure you want to continue?
-      </p>
+      this.props.service.getId() !== this.state.serviceNameConfirmationValue
     );
   }
 
@@ -140,42 +133,142 @@ class ServiceDestroyModal extends React.Component {
     }, REDIRECT_DELAY);
   }
 
-  render() {
-    const { isPending, onClose, open, service } = this.props;
+  getCloseButton() {
+    return (
+      <div className="row text-align-center">
+        <button
+          className="button button-primary button-medium"
+          onClick={this.handleModalClose}
+        >
+          Close
+        </button>
+      </div>
+    );
+  }
 
-    let itemText = "Service";
+  getDestroyFrameworkModal() {
+    const { open, service, intl } = this.props;
+    const serviceName = service.getId();
+
+    return (
+      <Modal
+        header={this.getModalHeading()}
+        footer={this.getCloseButton()}
+        modalClass="modal"
+        onClose={this.handleModalClose}
+        open={open}
+        showHeader={true}
+        showFooter={true}
+        subHeader={this.getSubHeader()}
+      >
+        <p>
+          {intl.formatMessage({
+            id: "SERVICE_ACTIONS.DELETE_SERVICE_FRAMEWORK"
+          })}
+          <a
+            href="https://docs.mesosphere.com/service-docs/"
+            target="_blank"
+            title={intl.formatMessage({
+              id: "COMMON.DOCUMENTATION_TITLE"
+            })}
+          >
+            {intl.formatMessage({
+              id: "COMMON.DOCUMENTATION"
+            })}
+          </a>
+          {intl.formatMessage({
+            id: "SERVICE_ACTIONS.DELETE_SERVICE_FRAMEWORK_2"
+          })}
+        </p>
+        <div className="flush-top snippet-wrapper">
+          <ClickToSelect>
+            <pre className="prettyprint flush-bottom">
+              dcos package uninstall
+              {" "}
+              {serviceName.replace("/", "")} --app-id={serviceName}
+            </pre>
+          </ClickToSelect>
+        </div>
+        {this.getErrorMessage()}
+      </Modal>
+    );
+  }
+
+  getDestroyServiceModal() {
+    const { open, service, intl } = this.props;
+    const serviceName = service.getId();
+
+    let itemText = `${StringUtil.capitalize(UserActions.DELETE)}`;
 
     if (service instanceof Pod) {
-      itemText = "Pod";
+      itemText += " Pod";
     }
 
     if (service instanceof ServiceTree) {
-      itemText = "Group";
+      itemText += " Group";
     }
-
-    const heading = (
-      <ModalHeading className="text-danger">
-        {StringUtil.capitalize(UserActions.DELETE)} {itemText}
-      </ModalHeading>
-    );
 
     return (
       <Confirm
-        disabled={isPending}
-        header={heading}
+        disabled={this.getIsRightButtonDisabled()}
+        header={this.getModalHeading()}
         open={open}
-        onClose={onClose}
+        onClose={this.handleModalClose}
         leftButtonText="Cancel"
-        leftButtonCallback={onClose}
-        rightButtonText={`${StringUtil.capitalize(UserActions.DELETE)} ${itemText}`}
+        leftButtonCallback={this.handleModalClose}
+        rightButtonText={itemText}
         rightButtonClassName="button button-danger"
         rightButtonCallback={this.handleRightButtonClick}
         showHeader={true}
       >
-        {this.getDestroyMessage()}
+        <p>
+          {intl.formatMessage({
+            id: "SERVICE_ACTIONS.DELETE_SERVICE"
+          })}
+          <strong>{serviceName}</strong>
+          {intl.formatMessage({
+            id: "SERVICE_ACTIONS.DELETE_SERVICE_2"
+          })}
+        </p>
+        <input
+          className="form-control filter-input-text"
+          onChange={this.handleChangeInputFieldDestroy}
+          type="text"
+          value={this.state.serviceNameConfirmationValue}
+        />
         {this.getErrorMessage()}
       </Confirm>
     );
+  }
+
+  getModalHeading() {
+    return (
+      <ModalHeading className="text-danger">
+        {StringUtil.capitalize(UserActions.DELETE)}
+      </ModalHeading>
+    );
+  }
+
+  getSubHeader() {
+    if (!this.props.subHeaderContent) {
+      return null;
+    }
+
+    return (
+      <p className="text-align-center flush-bottom">
+        {this.props.subHeaderContent}
+      </p>
+    );
+  }
+
+  render() {
+    const { service } = this.props;
+
+    if (service instanceof Framework) {
+      return this.getDestroyFrameworkModal();
+    }
+
+    return this.getDestroyServiceModal();
   }
 }
 
@@ -186,13 +279,16 @@ ServiceDestroyModal.contextTypes = {
 ServiceDestroyModal.propTypes = {
   deleteItem: PropTypes.func.isRequired,
   errors: PropTypes.oneOfType([PropTypes.object, PropTypes.string]),
+  intl: intlShape.isRequired,
   isPending: PropTypes.bool.isRequired,
   onClose: PropTypes.func.isRequired,
   open: PropTypes.bool.isRequired,
   service: PropTypes.oneOfType([
+    PropTypes.instanceOf(Framework),
+    PropTypes.instanceOf(Pod),
     PropTypes.instanceOf(ServiceTree),
     PropTypes.instanceOf(Service)
   ]).isRequired
 };
 
-module.exports = ServiceDestroyModal;
+module.exports = injectIntl(ServiceDestroyModal);
